# Optimize reconstructing of parity fragments