--- a/liberasurecode.patch
+++ b/liberasurecode.patch
@@ -11,7 +11,7 @@ index 52554a9..0a6863d 100644
    int *tbl1_r;
    int *tbl2_l;
 diff --git a/src/backends/jerasure/jerasure_rs_cauchy.c b/src/backends/jerasure/jerasure_rs_cauchy.c
-index 82d796a..12323c8 100644
+index 82d796a..9628314 100644
 --- a/src/backends/jerasure/jerasure_rs_cauchy.c
 +++ b/src/backends/jerasure/jerasure_rs_cauchy.c
 @@ -28,6 +28,8 @@
@@ -23,7 +23,17 @@ index 82d796a..12323c8 100644
  
  #include "erasurecode.h"
  #include "erasurecode_backend.h"
-@@ -147,6 +149,7 @@ static int jerasure_rs_cauchy_reconstruct(void *desc, char **data, char **parity
+@@ -104,6 +106,9 @@ struct jerasure_rs_cauchy_descriptor {
+ static void free_rs_cauchy_desc(
+         struct jerasure_rs_cauchy_descriptor *jerasure_desc );
+ 
++int *jerasure_bitmatrix_multiply(int *m1, int *m2, int r1, int c1,
++                                 int r2, int c2);
++
+ 
+ static int jerasure_rs_cauchy_encode(void *desc, char **data, char **parity,
+         int blocksize)
+@@ -147,6 +152,7 @@ static int jerasure_rs_cauchy_reconstruct(void *desc, char **data, char **parity
      int *erased = NULL;           /* k+m length list of erased frag ids */
      int *dm_ids = NULL;           /* k length list of fragment ids */
      int *decoding_matrix = NULL;  /* matrix for decoding */
@@ -31,7 +41,7 @@ index 82d796a..12323c8 100644
  
      struct jerasure_rs_cauchy_descriptor *jerasure_desc = 
          (struct jerasure_rs_cauchy_descriptor*) desc;
-@@ -180,26 +183,38 @@ static int jerasure_rs_cauchy_reconstruct(void *desc, char **data, char **parity
+@@ -180,26 +186,37 @@ static int jerasure_rs_cauchy_reconstruct(void *desc, char **data, char **parity
              goto out;
          }
      } else {
@@ -65,8 +75,7 @@ index 82d796a..12323c8 100644
 +          goto out;
 +        }
 +
-+        // Because in the patch jerasure is statically linked, this symbol is visible.
-+        decoding_row_prod = jerasure_matrix_multiply(decoding_row_orig, decoding_matrix, w, k * w, k * w, k * w, 1);
++        decoding_row_prod = jerasure_bitmatrix_multiply(decoding_row_orig, decoding_matrix, w, k * w, k * w, k * w);
 +        if (decoding_row_prod == NULL) {
 +          ret = -1;
 +          goto out;
@@ -84,7 +93,7 @@ index 82d796a..12323c8 100644
      
      return ret;
  }
-@@ -268,87 +283,16 @@ static void * jerasure_rs_cauchy_init(struct ec_backend_args *args,
+@@ -268,87 +285,16 @@ static void * jerasure_rs_cauchy_init(struct ec_backend_args *args,
          }
      }
  
@@ -181,6 +190,35 @@ index 82d796a..12323c8 100644
  
      /* setup the Cauchy matrices and schedules */
      desc->matrix = desc->cauchy_original_coding_matrix(k, m, w);
+@@ -474,3 +420,28 @@ struct ec_backend_common backend_jerasure_rs_cauchy = {
+                                            JERASURE_RS_CAUCHY_LIB_MINOR,
+                                            JERASURE_RS_CAUCHY_LIB_REV),
+ };
++
++// bitmatrix's counterpart of jerasure_matrix_multiply.
++// We implement this function because we don't want to depend on gf-complete, and therefore introduce unnecessary race conditions.
++// Semantically equivalent to jerasure_matrix_multiply(m1, m2, r1, c1, r2, c2, 1).
++int *jerasure_bitmatrix_multiply(int *m1, int *m2, int r1, int c1,
++                                 int r2, int c2)
++{
++  int *product, i, j, k;
++
++  product = (int *) malloc(sizeof(int)*r1*c2);
++  if (product == NULL) {
++    return NULL;
++  }
++
++  for (i = 0; i < r1*c2; i++) product[i] = 0;
++
++  for (i = 0; i < r1; i++) {
++    for (j = 0; j < c2; j++) {
++      for (k = 0; k < r2; k++) {
++        product[i*c2+j] ^= m1[i*c1+k] && m2[k*c2+j];
++      }
++    }
++  }
++  return product;
++}
 diff --git a/src/backends/jerasure/jerasure_rs_vand.c b/src/backends/jerasure/jerasure_rs_vand.c
 index 9395046..143e00f 100644
 --- a/src/backends/jerasure/jerasure_rs_vand.c