@koba-e964 2019-08-05
note: make_decoding_matrix を呼ぶところは data フラグメントの reconstruct で使われるコードをそのまま持ってきたため、ほぼ問題ないと思われるが、生きているフラグメントが全て data である場合については念のためチェックする必要がある。